backend/requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database and ORM
supabase==2.0.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1

# AI and ML (Anthropic Claude)
# openai==1.3.7  # Commented out - using Anthropic instead
numpy==1.24.3
scikit-learn==1.3.2

# Web Scraping and Content Processing
beautifulsoup4==4.12.2
playwright==1.40.0
requests==2.31.0
aiohttp==3.9.1
lxml==4.9.3

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3

# YouTube Processing
youtube-transcript-api==0.6.1

# Data Processing
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
click==8.1.7
rich==13.7.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Monitoring and Logging
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Caching (for Phase 7)
redis==5.0.1

# Security
cryptography==41.0.8
anthropic==0.64.0
openai==1.3.7
